fix pdf converter, now can skip previosly downloaded

+++ download_studybook.py
@@ -4,6 +4,7 @@ from time import sleep
 from bs4 import BeautifulSoup
 from urllib.parse import urljoin
 import requests
+import os
 
 
 def download_from_the_internet(url):
@@ -29,19 +30,22 @@ image_url = 'https://reader.lecta.rosuchebnik.ru/8233-63/data/'
 template_image = "temp{}.png"
 
 for i in range(1, 370):
-    print(i)
-    current_url = url.format(make_inset(i))
-    
-    html = download_from_the_internet(current_url)
-    image_url = urljoin(image_url, parse_html(html))
-    print(image_url)
-    
-    response = requests.get(image_url, stream=True)
-    image_data = response.content
     temp_image = template_image.format(i)
-    with open(temp_image, "wb") as f:
-        f.write(image_data)
+    if not os.path.exists(temp_image):
+        print(i)
+        current_url = url.format(make_inset(i))
+        
+        html = download_from_the_internet(current_url)
+        image_url = urljoin(image_url, parse_html(html))
+        print(image_url)
+        
+        response = requests.get(image_url, stream=True)
+        image_data = response.content
+        temp_image = template_image.format(i)
+        with open(temp_image, "wb") as f:
+            f.write(image_data)
+        
+        sleep(0.2)        
     pdf.add_page()
     pdf.image(temp_image)
-    sleep(0.2)
 pdf.output("algebra8.pdf", "F")
